refactor: 다리 지도 클래스 분리
